Safari-609.1.11.1/LayoutTests/fast/dom/SelectorAPI/matches-definition.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
</head>
<body>
</body>
<script>
description('Test the definition of Element.matches().');

shouldBeUndefined('Element.matches');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "matches").enumerable');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "matches").writable');
shouldBe('Element.prototype.matches.length', '1');

shouldBeUndefined('Element.webkitMatchesSelector');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "webkitMatchesSelector").enumerable');
shouldBeTrue('Object.getOwnPropertyDescriptor(Element.prototype, "webkitMatchesSelector").writable');
shouldBe('Element.prototype.webkitMatchesSelector.length', '1');
</script>
<script src="../../../resources/js-test-post.js"></script>
</html>
