LayoutTests/fast/dom/SelectorAPI/matches-definition-expected.txt

Test the definition of Element.matches().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Element.matches is undefined.
PASS Object.getOwnPropertyDescriptor(Element.prototype, "matches").enumerable is true
PASS Object.getOwnPropertyDescriptor(Element.prototype, "matches").writable is true
PASS Element.prototype.matches.length is 1
PASS Element.webkitMatchesSelector is undefined.
PASS Object.getOwnPropertyDescriptor(Element.prototype, "webkitMatchesSelector").enumerable is true
PASS Object.getOwnPropertyDescriptor(Element.prototype, "webkitMatchesSelector").writable is true
PASS Element.prototype.webkitMatchesSelector.length is 1
PASS successfullyParsed is true

TEST COMPLETE
